# docs/relay/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UAM Relay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agents/{address}/public-key": {
      "get": {
        "summary": "Get Public Key",
        "description": "Return the public key for a registered agent.\n\nUnauthenticated -- any caller can look up any agent's public key.\nThis is required for encrypting the first message (handshake.request)\nwhen the sender has no prior relationship with the recipient.",
        "operationId": "get_public_key_api_v1_agents__address__public_key_get",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{address}": {
      "patch": {
        "summary": "Patch Agent",
        "description": "Update agent fields (display_name, contact_card, public_key).\n\nRequires Bearer token auth.  Agent can only update their own record.\nIf ``public_key`` is provided, it must be a valid Ed25519 key.",
        "operationId": "patch_agent_api_v1_agents__address__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Agent",
        "description": "Soft-deactivate an agent.\n\nRequires Bearer token auth.  Agent can only deactivate themselves.",
        "operationId": "delete_agent_api_v1_agents__address__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Agent Api V1 Agents  Address  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{address}/reactivate": {
      "post": {
        "summary": "Reactivate Agent Endpoint",
        "description": "Reactivate a soft-deleted agent.\n\nRequires Bearer token auth.  The token must match the original agent's\ntoken (auth does not filter by deleted_at so soft-deleted agents can\nstill authenticate for reactivation).",
        "operationId": "reactivate_agent_endpoint_api_v1_agents__address__reactivate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/send": {
      "post": {
        "summary": "Demo Send",
        "description": "Send a message from the demo session to any registered agent.\n\nThe relay signs and encrypts the envelope on behalf of the ephemeral\nagent using the server-held private key.",
        "operationId": "demo_send_api_v1_demo_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/inbox": {
      "get": {
        "summary": "Demo Inbox",
        "description": "Return decrypted plaintext messages for the demo session.\n\nThe relay decrypts each stored envelope using the server-held private\nkey, so the browser receives plaintext content.",
        "operationId": "demo_inbox_api_v1_demo_inbox_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Demo session ID",
              "title": "Session Id"
            },
            "description": "Demo session ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoInboxResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "public_key": {
            "type": "string",
            "title": "Public Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "address",
          "public_key",
          "status",
          "created_at"
        ],
        "title": "AgentResponse"
      },
...
      "DemoInboxResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/DemoMessage"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "DemoInboxResponse"
      },
      "DemoMessage": {
        "properties": {
          "from_address": {
            "type": "string",
            "title": "From Address"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          }
        },
        "type": "object",
        "required": [
          "from_address",
          "content",
          "timestamp",
          "message_id"
        ],
        "title": "DemoMessage"
      },
      "DemoSendRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "to_address": {
            "type": "string",
            "title": "To Address"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "to_address",
          "message"
        ],
        "title": "DemoSendRequest"
      },
      "DemoSendResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          }
        },
        "type": "object",
        "required": [
          "message_id"
        ],
        "title": "DemoSendResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicKeyResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "public_key": {
            "type": "string",
            "title": "Public Key"
          },
          "tier": {
            "type": "integer",
            "title": "Tier",
            "default": 1
          },
          "verified_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Domain"
          }
        },
        "type": "object",
        "required": [
          "address",
          "public_key"
        ],
        "title": "PublicKeyResponse"
      },
...
      "UpdateAgentRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "contact_card": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Card"
          },
          "public_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Key"
          }
        },
        "type": "object",
        "title": "UpdateAgentRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}